% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards.R
\name{add_ards}
\alias{add_ards}
\title{Adds data to an Analysis Results Dataset}
\usage{
add_ards(
  data,
  statvars,
  statdesc = NULL,
  byvars = NULL,
  trtvar = NULL,
  paramcd = NULL,
  anal_var = NULL,
  anal_val = NULL
)
}
\arguments{
\item{data}{The input dataset to create analysis results for.}

\item{statvars}{A vector of column names that identify the desired results.
Statvar columns must be numeric.  This parameter is required.}

\item{statdesc}{A vector of column names that identify a description value
for each statvar.}

\item{byvars}{A vector of column names to use for by variables.}

\item{trtvar}{A column name to use for the treatment variable.}

\item{paramcd}{A description of the analysis parameter.}

\item{anal_var}{A column name for the analysis variable.}

\item{anal_val}{The analysis variable value.  Can be identified by a column
name or a vector of values.}
}
\value{
The input data frame, unaltered.
}
\description{
The \code{add_ards} function dumps data from an input dataset
to the ARDS dataset.  The function is designed to be pipe-friendly, and will
return the input dataset unaltered.  The parameters on the function
define how to extract the desired data from the input dataset.
The "statvars" parameter defines which column names contain desired
analysis data.  The values in these columns will be used to populate the
"statval" variable in the output dataset.  Other parameters are used to
define identifying information for the statistics value, and are optional.
}
