% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards.R
\name{get_ards}
\alias{get_ards}
\title{Returns the Current Analysis Results Dataset}
\usage{
get_ards()
}
\value{
A data frame of the current analysis results.
}
\description{
The \code{get_ards} function returns the current state
of the analysis dataset.  This data frame may be saved to disk, saved in
a database, or examined from code.  The function takes no parameters.
}
