% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ards.R
\name{init_ards}
\alias{init_ards}
\title{Initialize the Analysis Results Dataset}
\usage{
init_ards(
  studyid = NULL,
  tableid = NULL,
  adsns = NULL,
  population = NULL,
  time = NULL,
  where = NULL,
  reset = TRUE
)
}
\arguments{
\item{studyid}{The study for which the analysis was performed. Optional string value.}

\item{tableid}{A table identifier to use for the results. Optional string
value.}

\item{adsns}{A vector of source dataset names.}

\item{population}{A description of the analysis population.}

\item{time}{A description of the time frame used in the analysis.}

\item{where}{A description of the criteria used to subset the data for analysis.}

\item{reset}{If true, clear out existing ARDS dataset and replace with
empty template.  Otherwise, just assign parameters to template.  The default
value is TRUE.}
}
\value{
The initialized analysis dataset.
}
\description{
A function to initialize the Analysis Results Dataset (ARDS).
This function will
first create a data template in the desired structure, and then
populates common values across the dataset.  These common values will be
repeated on each row of the analysis data frame for subsequent inserts.
Use the "reset" parameter to clear out any existing data and start a new
analysis dataset.
}
